fix project configuration
