--- company-site/src/components/UI/News.jsx
@@ -1,3 +1,0 @@
-export const News = (props) => {
-
-}
