+++ company-site/src/components/landing/index.js
@@ -1,0 +1,15 @@
+import { About } from "./about";
+import { Contacts } from "./contacts";
+import { Footer } from "./footer";
+import { Header } from "./header";
+import { Info } from "./info";
+import { Supports } from "./supports";
+
+export {
+    About,
+    Contacts,
+    Footer,
+    Header,
+    Info,
+    Supports
+}
