+++ company-site/src/style/info.css
