--- hakaton_experience/style.css
@@ -1,381 +1,0 @@
-*
-{
-  margin: 0;
-  padding: 0;
-  box-sizing: border-box;
-  font-family:Arial, Helvetica, sans-serif;
-}
-
-.section{
-  width: 100%;
-  min-height: 100vh;
-  padding: 100px;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  background: #fff;
-  padding-bottom: 0;
-}
-
-.section-company{
-  width: 100%;
-  padding-left: 100px;
-  padding-right: 100px;
-  margin-top: 10px;
-  margin-bottom: 125px;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  background: #fff;
-  padding-bottom: 0;
-}
-
-.header{
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 100%;
-  padding: 20px 100px;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-}
-
-.logo{
-  position: relative;
-  max-width: 190px;
-  margin-left: 5px;
-}
-
-.header-nav{
-  position: relative;
-  display: flex;
-}
-
-.header-nav-li{
-  list-style: none;
-}
-
-.contact-inf{
-  width: 250px; 
-  height: 60px; 
-  background: black; 
-  border-radius: 10px;
-  list-style: none;
-  color: white; 
-  
-}
-
-
-header ul li a{
-  display: inline-block;
-  color:  #29272A;
-  font-weight: 400;
-  margin-left: 40px;
-  text-decoration: none;
-}
-
-.content{
-  position: relative;
-  width: 100%;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-}
-
-.content-textBox{
-  position: relative;
-  max-width: 700px;
-  font-size: 20px;
-}
-
-.content-textBox-h2{
-  color: #0B132A;
-  font-size: 50px;
-  line-height: 1.4em;
-  font-weight: bold;
-}
-
-.content-textBox-h3{
-  color: #0B132A;
-  font-size: 30px;
-  line-height: 1.4em;
-  font-weight: bold;
-  text-align: center;
-  padding-bottom: 27px;
-}
-.content-textBox-p{
-  color: #29272A;
-  margin-top: 35px;
-  font-size: 24px;
-  line-height: 37.50px;
-  letter-spacing: 0.50px;
-  word-wrap: break-word;
-  text-align: justify;
-}
-
-.goods{
-  padding: 0%;
-}
-
-.section-second{
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    text-align: center;
-}
-
-.shadow-button{
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    text-align: center;
-    height: 300px;
-    width: 90%;
-    background: rgb(117, 117, 117, 0.03);
-    border-radius: 30px;
-}
-
-.buttons{
-    display: flex;
-    position: relative;
-    justify-content: center;
-    opacity: 100%;
-    align-items: center;
-    text-align: center;
-    border-radius: 10px;
-    height: 179px;
-    width: 1140px;
-}
-
-.button-left,
-.button-right{
-    display: flex;
-    height: 179px;
-    width: 570px;
-    background: #fff;
-    align-items: center;
-    justify-content: center;
-}
-
-.text-in-button{
-    width: 398px;
-    height: 62.65px;
-    flex-shrink: 0;
-    color: #000;
-    text-align: center;
-    font-size: 32px;
-    font-style: normal;
-    font-weight: 400;
-    line-height: 35px; /* 109.375% */
-    align-items: center;
-}
-
-.stick{
-    width: 2px;
-    height: 112.388px;
-    flex-shrink: 0;
-    stroke-width: 2px;
-    background: rgb(117, 117, 117, 0.3);
-}
-
-.laptop-box{
-    display: flex;
-    width: 667px;
-    height: 532px;
-    flex-shrink: 0;
-    margin-left: 70px;
-}
-.laptop{
-    display: flex;
-    width: 667px;
-    height: 532px;
-    flex-shrink: 0;
-
-    /* background: #000; */
-}
-
-.basic-services{
-    display: flex;
-    position: relative;
-    width: 100%;
-
-}
-
-.services{
-    position: relative;
-    margin-left: 8%;
-    width: 570px;
-}
-
-.heading-services{
-    display: flex;
-    position: relative;
-    color: #0B132A;
-    margin-top: 10%;
-    font-size: 35px;
-    font-style: normal;
-    font-weight: 500;
-    line-height: 50px; /* 142.857% */
-}
-
-.services .text-services{
-    display: flex;
-    position: relative;
-    color: #4F5665;
-    font-size: 16px;
-    font-style: normal;
-    font-weight: 400;
-    line-height: 30px; /* 187.5% */
-    width: 412px;
-    margin-top: 24px;
-}
-
-.dots-list{
-    margin-top: 51px;
-    position: relative;
-}
-.dot{
-    margin-top: 15px;
-    display: flex;
-    position: relative;
-}
-
-.dots-text{
-    margin-left: 10px;
-    color: #4F5665;
-    font-size: 14px;
-    font-style: normal;
-    font-weight: 400;
-    line-height: 30px; /* 214.286% */
-}
-
-.team{
-    /* display: flex; */
-    position: relative;
-    width: 100%;
-}
-.name-team{
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    color: #0B132A;
-    font-size: 35px;
-    font-style: normal;
-    font-weight: 500;
-    line-height: 50px; /* 142.857% */
-}
-
-.line{
-    display: flex;
-    position: relative;
-    justify-content: center;
-    align-items: center;
-    grid-column-gap: 100px;
-}
-
-.grid{
-  margin-bottom: 100px;
-}
-
-.person{
-    margin-top: 86px;    
-}
-
-.person-image{
-    width: 240px;
-    height: 240px;
-    flex-shrink: 0;
-    border-radius: 50%;
-    background: #4F5665;
-}
-
-.person p{
-    margin-top: 15px;
-    width: 222.051px;
-    height: 51.671px;
-    flex-shrink: 0;
-    color: #000;
-    text-align: center;
-    font-size: 20px;
-    font-style: normal;
-    font-weight: 400;
-    line-height: 30px; /* 150% */
-}
-
-
-.section-inf{
-  position: relative;
-  display: flex;
-  justify-content: center;
-  width: 100%;
-  align-items: center;
-  background: #fff;
-  padding-top: 0 px;
-}
-
-.content-inf{
-  padding-top: 0px;
-  white-space: normal;  
-  vertical-align: top;
-  display: grid;
-}
-
-.content-textBox-inf{
-  color: #29272A;
-  padding-left: 30px;
-  font-size: 16px;
-  line-height: 20px;
-  letter-spacing: 0.50px;
-  word-wrap: break-word;
-  text-align: justify;
-}
-
-.map{
-    display: flex;
-    justify-content: center;
-    margin: 0;
-    padding: 0;
-    align-items: center;
-}
-.map-inf{
-  color: black;
-  font-size: 15px;
-  font-weight: 300;
-  word-wrap: break-word;
-  line-height: 1.4em;
-  text-align: center;
-  margin-top: 20px;
-}
-
-.info{
-  width: 100%;
-  display: flex;
-  line-height: 2em;
-  flex-direction: column;
-  padding: 75px;
-
-
-  font-size: 20px;
-}
-
-.content-footer-h3{
-  display: flex;
-  justify-content: center;
-}
-
-.content-footer{
-  display: flex;
-  padding-top: 0px;
-  white-space: normal;  
-  vertical-align: top;
-  align-items: center;
-  display: grid;
-}
-
-.content-textBox-left,
-.content-textBox-right{
-  display: flex;
-  position: relative;
-
-}
